requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
requests>=2.31.0
websockets>=12.0
httpx>=0.27.0
openai>=1.30.0
openenv-core>=0.1.0
